Me gusta y deseo

+++ Presentation/ShoppingCar/ShoppingCar.php
@@ -5,15 +5,24 @@
         <meta charset="UTF-8">
         <title></title>
         <script src="../../JS/jquery-3.1.1.min.js" type="text/javascript"></script>
+        
         <?php
-            include '../../Business/Account/AccountBusiness.php';
-            $accountBusiness = new AccountBusiness();
-            $result = $accountBusiness->getAllAccountAssetsBusiness();
+        include '../../Business/Account/AccountBusiness.php';
+        include_once '../../Business/DesiredProductBusiness/DesiredProductBusiness.php';
+        $accountBusiness = new AccountBusiness();
+        $result = $accountBusiness->getAllAccountAssetsBusiness();
+        
+
+        $desiredBusiness = new DesiredProductBusiness();
+        $resultDesired = $desiredBusiness->getDesiredProducts($_SESSION['idUser']);
+        if ($resultDesired != 0) {
+            $_SESSION['desired'] = $resultDesired;
+        }
         ?>
     </head>
     <body>
         <?php
-        if(!isset($_SESSION["idUser"])){
+        if (!isset($_SESSION["idUser"])) {
             session_start();
         }
         ?>
@@ -22,10 +31,12 @@
         <table>
             <tr>
                 <td><a href="../Modules/ClientView.php">Atrás</a></td>
+                <td><a href="./ShoppingCar.php?option=desired">Deseos</a></td>
+                <td><a href="./ShoppingCar.php?option=carrito">Carrito</a></td>
             </tr>
         </table>
         <hr>
-        <h1>Carrito compras</h1>
+        <h1>Productos agregados</h1>
         <br>        
         <table>
             <th>Nombre</th>
@@ -35,11 +46,12 @@
             <th>Precio</th> 
 
             <?php
+            $option = $_GET['option'];
             $total = 0;
-            if (sizeof($_SESSION['carrito']) > 0) {
-                for ($i = 0; $i < sizeof($_SESSION['carrito']); $i++) {
+            if (sizeof($_SESSION[$option]) > 0) {
+                for ($i = 0; $i < sizeof($_SESSION[$option]); $i++) {
 
-                    $product = $_SESSION['carrito'][$i];
+                    $product = $_SESSION[$option][$i];
                     $infoProduct = split(";", $product);
                     $idProduct = $infoProduct[0];
                     ?>
@@ -59,8 +71,20 @@
                             }
                         }
                         ?>
-                        <td><input type="submit" class="btnDelete" id="btnDelete" value="Quitar"
-                                   onclick="return deleteItem('<?php echo $idProduct; ?>')"/></td>
+                        <?php
+                        if ($option == "carrito") {
+                            ?>
+                            <td><input type="submit" class="btnDelete" id="btnDelete" value="Quitar"
+                                       onclick="return deleteItem('<?php echo $idProduct; ?>')"/></td>
+                                <?php
+                            } else {
+                                ?>
+                            <td><input type="submit" class="btnDeleteDesired" id="btnDeleteDesired" value="Quitar"
+                                       onclick="return deleteDesiredProduct('<?php echo $idProduct; ?>')"/></td>
+                                <?php
+                            }
+                            ?>        
+
                     </tr>
 
 
@@ -68,7 +92,7 @@
                 }
             } else {
                 ?>
-                <tr><td><label>El carrito está vacío</label></td></tr>
+                <tr><td><label>No hay productos agregados</label></td></tr>
                 <?php
             }
             ?>
@@ -82,14 +106,14 @@
                         <!-- Cuenta -->
                         <label>Cuenta: </label>
                         <select name="account">
-                        <?php
-                        foreach ($result as $tem) {
-                            echo "<option value='" .$tem->idAccount ."'>". $tem->cardNumber ."</option>";
-                        }
-                        ?>
+                            <?php
+                            foreach ($result as $tem) {
+                                echo "<option value='" . $tem->idAccount . "'>" . $tem->cardNumber . "</option>";
+                            }
+                            ?>
                         </select>
                         <!-- Fin de cuenta -->
-                        
+
                         <input type="submit" value="Comprar" id=""/>
                     </form>
 
@@ -100,7 +124,7 @@
         <label id="txtMessage"></label>
     </center>
 </body>
-<?php 
+<?php
 if (isset($_GET['success'])) {
     echo '<script>                        
              document.getElementById("txtMessage").innerHTML = "Compra con éxito";
@@ -109,7 +133,7 @@ if (isset($_GET['success'])) {
     echo '<script>                
               document.getElementById("txtMessage").innerHTML = "Compra fallida";
           </script>';
-}else if (isset($_GET['errorData'])) {
+} else if (isset($_GET['errorData'])) {
     echo '<script>                
               document.getElementById("txtMessage").innerHTML = "El carrito está vacío, o seleccione una cuenta";
           </script>';
@@ -130,7 +154,7 @@ if (isset($_GET['success'])) {
 
             },
             success: function (response) {
-                window.location = "./ShoppingCar.php";
+                window.location = "./ShoppingCar.php?option=carrito";
             },
             error: function () {
                 document.getElementById('txtMessage').innerHTML = "Error al eliminar";
@@ -138,8 +162,31 @@ if (isset($_GET['success'])) {
         });
 
     }
-
 </script>
+
+<script>
+    function deleteDesiredProduct(idProduct) {
+        var parametros = {
+            "idProduct": idProduct
+        };
+        $.ajax({
+            data: parametros,
+            url: '../../Business/DesiredProductBusiness/DesiredProductUpdate.php',
+            type: 'post',
+            beforeSend: function () {
+
+            },
+            success: function (response) {
+                window.location = "./ShoppingCar.php?option=desired";
+            },
+            error: function () {
+                document.getElementById('txtMessage').innerHTML = "Error al eliminar";
+            }
+        });
+
+    }
+</script>
+
